+++ src/Board320_240.h
@@ -80,6 +80,7 @@ public:
   //
   void initBoard() override;
   void afterSetup() override;
+  void boardLoop() override;
   void mainLoop() override;
   bool skipAdapterScan() override;
   void goToSleep();
@@ -104,6 +105,7 @@ public:
   void turnOffScreen() override;
   void setBrightness() override;
   void displayMessage(const char *row1, const char *row2) override;
+  bool confirmMessage(const char *row1, const char *row2) override;
   void redrawScreen() override;
   // Custom screens
   void drawBigCell(int32_t x, int32_t y, int32_t w, int32_t h, const char *text, const char *desc, uint16_t bgColor, uint16_t fgColor);