+++ src/BoardM5stackCore2.h
@@ -14,10 +14,13 @@ class BoardM5stackCore2 : public Board320_240 {
     void initBoard() override;
     void afterSetup() override;
     void wakeupBoard() override;
+    void boardLoop() override;
     void mainLoop() override;
     bool isButtonPressed(int button) override;
     void enterSleepMode(int secs) override;
     void Write1Byte(uint8_t Addr, uint8_t Data);
     uint8_t Read8bit(uint8_t Addr);
     static void eventDisplay(Event& e);
+    void setTime(String timestamp) override;
+    void ntpSync() override;
 };