+++ src/menu.h
@@ -2,7 +2,7 @@
 
 #include "config.h"
 
-#define MENU_SIZE 112
+#define MENU_SIZE 115
 
 
 MENU_ITEM menuItemsSource[MENU_SIZE] = {
@@ -16,6 +16,7 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_FACTORY_RESET,            MENU_TOP_LEVEL,     MENU_NO_MENU,     "Factory reset"},
     {MENU_SAVE_SETTINGS,            MENU_TOP_LEVEL,     MENU_NO_MENU,     "Save settings"},
     {MENU_APP_VERSION,              MENU_TOP_LEVEL,     MENU_NO_MENU,     "OTA update, curr:"},
+    {MENU_MEMORY_USAGE,             MENU_TOP_LEVEL,     MENU_NO_MENU,     "Memory usage"},
     {MENU_SHUTDOWN,                 MENU_TOP_LEVEL,     MENU_NO_MENU,     "Shutdown"},
     {MENU_CAR_COMMANDS,             MENU_TOP_LEVEL,     MENU_NO_MENU,     "Car commands [CAN]"},
 
@@ -42,9 +43,6 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_ADAPTER_BT3,              MENU_ADAPTER_TYPE,  MENU_NO_MENU,     "Bluetooth3 serial"},
     {MENU_ADAPTER_LOAD_TEST_DATA,              MENU_ADAPTER_TYPE,  MENU_NO_MENU,     "Load demo data"},
 
-    {FACTORY_RESET_TOP,             MENU_FACTORY_RESET, MENU_TOP_LEVEL,   "<- parent menu"},
-    {FACTORY_RESET_CONFIRM,         MENU_FACTORY_RESET, MENU_NO_MENU,     "Confirm factory reset"},
-
     {MENU_OTHER_TOP,                MENU_OTHERS,        MENU_TOP_LEVEL,   "<- parent menu"},
     {MENU_WIFI,                     MENU_OTHERS,        MENU_NO_MENU,     "WiFi network"},
     {MENU_SDCARD,                   MENU_OTHERS,        MENU_NO_MENU,     "SD card"},
@@ -53,7 +51,6 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_SERIAL_CONSOLE,           MENU_OTHERS,        MENU_NO_MENU,     "Serial console"},
     {MENU_VOLTMETER,                MENU_OTHERS,        MENU_NO_MENU,     "Voltmeter INA3221"},
     {MENU_DEBUG_LEVEL,              MENU_OTHERS,        MENU_NO_MENU,     "Debug level"},
-    //{MENU_NTP,                      MENU_OTHERS,        MENU_NO_MENU,     "[dev] NTP"},
     {MENU_SCREEN_ROTATION,          MENU_OTHERS,        MENU_NO_MENU,     "Screen rotation"},
     {MENU_DEFAULT_SCREEN,           MENU_OTHERS,        MENU_NO_MENU,     "Default screen"},
     {MENU_SCREEN_BRIGHTNESS,        MENU_OTHERS,        MENU_NO_MENU,     "LCD brightness"},
@@ -76,6 +73,7 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_WIFI_PASSWORD,            MENU_WIFI,          MENU_NO_MENU,     "Password:"},
     {MENU_WIFI_SSID2,               MENU_WIFI,          MENU_NO_MENU,     "SSID2:"},
     {MENU_WIFI_PASSWORD2,           MENU_WIFI,          MENU_NO_MENU,     "Password2:"},
+    {MENU_WIFI_NTP,                 MENU_WIFI,          MENU_NO_MENU,     "NTP sync:"},
     {MENU_WIFI_ACTIVE,              MENU_WIFI,          MENU_NO_MENU,     "Active:"},
     {MENU_WIFI_IPADDR,              MENU_WIFI,          MENU_NO_MENU,     "IP addr:"},
 
@@ -84,11 +82,14 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_SDCARD_AUTOSTARTLOG,      MENU_SDCARD,        MENU_NO_MENU,     "Autostart log enabled"},
     {MENU_SDCARD_MOUNT_STATUS,      MENU_SDCARD,        MENU_NO_MENU,     "Status"},
     {MENU_SDCARD_REC,               MENU_SDCARD,        MENU_NO_MENU,     "Record"},
+    {MENU_SDCARD_SETTINGS_SAVE,     MENU_SDCARD,        MENU_NO_MENU,     "Backup settings to SDCARD"},
+    {MENU_SDCARD_SETTINGS_RESTORE,  MENU_SDCARD,        MENU_NO_MENU,     "Restore settings from SD"},
 //  {MENU_SDCARD_INTERVAL,          MENU_SDCARD,        MENU_NO_MENU,     "Log interval sec."},
 
     {MENU_VOLTMETER_TOP,            MENU_VOLTMETER,     MENU_OTHERS,      "<- parent menu"},
     {MENU_VOLTMETER_ENABLED,        MENU_VOLTMETER,     MENU_NO_MENU,     "Voltmeter enabled"},
     {MENU_VOLTMETER_SLEEP,          MENU_VOLTMETER,     MENU_NO_MENU,     "Control SleepMode"},
+    {MENU_VOLTMETER_INFO,           MENU_VOLTMETER,     MENU_NO_MENU,     "Show voltage/current"},
     {MENU_VOLTMETER_SLEEPVOL,       MENU_VOLTMETER,     MENU_NO_MENU,     "Sleep Vol."},
     {MENU_VOLTMETER_WAKEUPVOL,      MENU_VOLTMETER,     MENU_NO_MENU,     "WakeUp Vol."},
     {MENU_VOLTMETER_CUTOFFVOL,      MENU_VOLTMETER,     MENU_NO_MENU,     "CutOff Vol."},